Revert "Load bundled GraalJS, give priority to GraalVM if present."
