--- ide/libs.graalsdk.system/manifest.mf
@@ -1,7 +1,0 @@
-Manifest-Version: 1.0
-AutoUpdate-Show-In-Client: false
-OpenIDE-Module: org.netbeans.libs.graalsdk.system
-OpenIDE-Module-Localizing-Bundle: org/netbeans/libs/graalsdk/system/Bundle.properties
-OpenIDE-Module-Specification-Version: 1.20
-OpenIDE-Module-Provides: org.netbeans.spi.scripting.EngineProvider
-OpenIDE-Module-Package-Dependencies: org.graalvm.polyglot[Engine]
