+++ ide/libs.graalsdk/manifest.mf
@@ -5,5 +5,3 @@ OpenIDE-Module-Localizing-Bundle: org/netbeans/libs/graalsdk/Bundle.properties
 OpenIDE-Module-Specification-Version: 1.20
 OpenIDE-Module-Provides: org.netbeans.spi.scripting.EngineProvider
 OpenIDE-Module-Recommends: com.oracle.truffle.polyglot.PolyglotImpl
-OpenIDE-Module-Hide-Classpath-Packages: org.graalvm.collections.**, org.graalvm.home.**, org.graalvm.nativeimage.**,
-    org.graalvm.options.**, org.graalvm.polyglot.**, org.graalvm.word.**
