+++ ide/libs.graalsdk/test/unit/src/org/netbeans/libs/graalsdk/GraalEnginesTest.java
@@ -16,11 +16,9 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.netbeans.libs.graalsdk.system;
+package org.netbeans.libs.graalsdk;
 
-import java.lang.reflect.Field;
 import java.util.ArrayList;
-import java.util.stream.Collectors;
 import java.util.List;
 import java.util.Map;
 import javax.script.Bindings;
@@ -29,20 +27,20 @@ import javax.script.ScriptContext;
 import javax.script.ScriptEngine;
 import javax.script.ScriptEngineFactory;
 import javax.script.ScriptEngineManager;
+import org.graalvm.polyglot.Context;
+import org.graalvm.polyglot.Value;
+import org.junit.Assert;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
 import org.junit.Assume;
 import static org.junit.Assume.assumeNotNull;
-import static org.junit.Assume.assumeTrue;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.netbeans.api.scripting.Scripting;
-import org.netbeans.junit.NbTestCase;
 
-public final class GraalEnginesTest2 extends NbTestCase {
-
-    public GraalEnginesTest2(String name) {
-        super(name);
-    }
-    
+public final class GraalEnginesTest {
     @BeforeClass
     public static void skipIfNoPolyglotFound() {
         try {
@@ -53,48 +51,9 @@ public final class GraalEnginesTest2 extends NbTestCase {
             Assume.assumeNoException("Error when initializing Engine", ex);
         }
     }
-    
-    /**
-     * Checks that in the presence of GraalVM JS implementation the JDK implementation wins in both
-     * enumeration AND extension registrations.
-     * Succeeds if just 1 JS is available (non-GraalVM JDKs)
-     */
-    public void testJDKImplementationFirstOnGraalVM() throws Exception {
-        ScriptEngineManager man = Scripting.createManager();
-        List<ScriptEngineFactory> factories = man.getEngineFactories().stream().filter(sf -> sf.getEngineName().equals("GraalVM:js")).collect(Collectors.toList());
-        assumeTrue("Need GraalVM and bundled implementation", factories.size() > 1);
-        
-        Object o = getEngineLanguage(factories.get(0));
-        assertTrue("GraalVM JDK JS implementation should take precedence", 
-                o.getClass().getProtectionDomain().getCodeSource().getLocation().toString().startsWith("jrt:/"));
-        ScriptEngine jsEngine = man.getEngineByExtension("js");
-        assertNotNull(jsEngine);
-        
-        Field f = jsEngine.getClass().getDeclaredField("factory");
-        f.setAccessible(true);
-        assertSame(factories.get(0), f.get(jsEngine));
-    }
-    
-    
-    /**
-     * Must access using reflection, as the production code is loaded by a proper module classloader, 
-     * while this test code is loaded by SystemClassLoader. Something could change in the testing infrastructure....
-     * 
-     * @param sf
-     * @return
-     * @throws Exception 
-     */
-    private static Object getEngineLanguage(ScriptEngineFactory sf) throws Exception {
-        if (!sf.getClass().getName().endsWith("GraalEngineFactory")) {
-            return null;
-        }
-        Field f = sf.getClass().getDeclaredField("language");
-        f.setAccessible(true);
-        return f.get(sf);
-    }
 
     @Test
-    public void testInvokeEngineViaGeneratedScriptEngine() throws Exception {
+    public void invokeEngineViaGeneratedScriptEngine() {
         ScriptEngineManager man = Scripting.createManager();
         ScriptEngine llvm = man.getEngineByName("GraalVM:llvm");
         assumeNotNull("Need llvm. Found: " + man.getEngineFactories(), llvm);
@@ -117,11 +76,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
             }
             if (types.contains("text/javascript")) {
                 if (factory.getEngineName().startsWith("GraalVM:")) {
-                    if (graalvmJsFactory != null) {
-                        Object l = getEngineLanguage(graalvmJsFactory);
-                        assertTrue("GraalVM JDK JS implementation should take precedence", 
-                                l.getClass().getProtectionDomain().getCodeSource().getLocation().toString().startsWith("jrt:/"));
-                    }
+                    assertNull("No previous generic GraalVM javascript factory: " + graalvmJsFactory, graalvmJsFactory);
                     graalvmJsFactory = factory;
                 } else if (!factory.getEngineName().equalsIgnoreCase("Oracle Nashorn")) {
                     assertNull("No previous javascript factory: " + jsFactory, jsFactory);
@@ -131,8 +86,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
         }
 
         assertNotNull("llvm factory found: " + log, llvmFactory);
-        // nashorn no longer exists...
-        // assertNotNull("js factory found: " + log, jsFactory);
+        assertNotNull("js factory found: " + log, jsFactory);
         assertNotNull("Generic GraalVM js factory found: " + log, graalvmJsFactory);
     }
 
@@ -142,24 +96,21 @@ public final class GraalEnginesTest2 extends NbTestCase {
         + "mul\n"
         + "\n";
 
-    /**
-    This test cannot work now, as it finds the lib.graalsdk's Context that has no connection to GraalVM JVM's Context that actually
-    serves Python.
-    public void testPythonDirect() throws Exception {
+    @Test
+    public void pythonDirect() throws Exception {
         assumeNotNull("Need python", Scripting.createManager().getEngineByMimeType("text/x-python"));
         final Context ctx = Context.newBuilder().allowAllAccess(true).build();
         Value mul = ctx.eval("python", MUL);
         Value fourtyTwo = mul.execute(6, 7);
         Assert.assertEquals("Fourty two", 42, fourtyTwo.asInt());
     }
-    */
 
     public static interface Mul {
         public int multiplyTwoNumbers(int x, int y);
     }
 
     @Test
-    public void testPythonFn() throws Exception {
+    public void pythonFn() throws Exception {
         ScriptEngine python = Scripting.createManager().getEngineByMimeType("text/x-python");
         assumeNotNull("Need python", python);
         Object rawMul = python.eval(MUL);
@@ -178,7 +129,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
     }
 
     @Test
-    public void testJavaScriptFn() throws Exception {
+    public void javaScriptFn() throws Exception {
         ScriptEngine js = Scripting.createManager().getEngineByName("GraalVM:js");
         Object rawFn = js.eval("(function (x, y) { return x * y; })");
 
@@ -190,7 +141,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
     }
 
     @Test
-    public void testPythonObjAccess() throws Exception {
+    public void pythonObjAccess() throws Exception {
         ScriptEngine python = Scripting.createManager().getEngineByMimeType("text/x-python");
         assumeNotNull("Need python", python);
         Object rawPoint = python.eval(
@@ -213,7 +164,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
     }
 
     @Test
-    public void testReturnArrayInPython() throws Exception {
+    public void returnArrayInPython() throws Exception {
         ScriptEngine python = Scripting.createManager().getEngineByMimeType("text/x-python");
         assumeNotNull("Need python", python);
         python.eval("\n"
@@ -247,7 +198,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
 
 
     @Test
-    public void testReturnArrayInJS() throws Exception {
+    public void returnArrayInJS() throws Exception {
         Assume.assumeFalse("Broken in GraalVM 20.3.0 fixed in GraalVM 21.1.0", "25.272-b10-jvmci-20.3-b06".equals(System.getProperty("java.vm.version")));
 
         ScriptEngine js = Scripting.createManager().getEngineByName("GraalVM:js");
@@ -265,9 +216,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
 
         ArrLike like = ((Invocable) js).getInterface(raw, ArrLike.class);
         assertNotNull("Array like " + like, like);
-        
-        // known bug in GraalJS 20.3.0
-        // assertEquals("Length of five", 5, like.length());
+        assertEquals("Length of five", 5, like.length());
 
         
         assertEquals(1, list.get(0));
@@ -282,7 +231,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
     }
 
     @Test
-    public void testReturnMapInJS() throws Exception {
+    public void returnMapInJS() throws Exception {
         ScriptEngine js = Scripting.createManager().getEngineByName("GraalVM:js");
         Object fn = js.eval("(function() {\n"
                 + "  return {\n"
@@ -308,7 +257,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
     }
 
     @Test
-    public void testAccessPolyglotBindings() throws Exception {
+    public void accessPolyglotBindings() throws Exception {
         ScriptEngineManager man = Scripting.createManager();
         ScriptEngine js = man.getEngineByName("GraalVM:js");
         ScriptEngine python = man.getEngineByName("GraalVM:python");
@@ -342,7 +291,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
     }
 
     @Test
-    public void testAccessPolyglotBindings2() throws Exception {
+    public void accessPolyglotBindings2() throws Exception {
         ScriptEngineManager man = Scripting.createManager();
         ScriptEngine python = man.getEngineByName("GraalVM:python");
         ScriptEngine js = man.getEngineByName("GraalVM:js");
@@ -379,13 +328,11 @@ public final class GraalEnginesTest2 extends NbTestCase {
      * @throws Exception 
      */
     @Test
-    public void testPolyglotBindingsAsAttributes() throws Exception {
+    public void polyglotBindingsAsAttributes() throws Exception {
         ScriptEngineManager man = Scripting.newBuilder().build();
 
         ScriptEngine snake = man.getEngineByName("GraalVM:python");
         ScriptEngine js = man.getEngineByName("GraalVM:js");
-        assumeNotNull("Need python", snake);
-        assumeNotNull("Need js", js);
         
         snake.getContext().setAttribute("preSnake", 1111, ScriptContext.GLOBAL_SCOPE);
         js.getContext().setAttribute("preJs", 2222, ScriptContext.GLOBAL_SCOPE);
@@ -409,7 +356,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
     }
 
     @Test
-    public void testHostAccessGlobalAttributeWorks() throws Exception {
+    public void hostAccessGlobalAttributeWorks() throws Exception {
         ScriptEngineManager man = Scripting.createManager();
         ScriptEngine js = man.getEngineByName("GraalVM:js");
         
@@ -420,7 +367,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
     }
 
     @Test
-    public void testAllAccessEnabledBuilder() throws Exception {
+    public void allAccessEnabledBuilder() throws Exception {
         ScriptEngineManager man = Scripting.newBuilder().allowAllAccess(true).build();
         ScriptEngine js = man.getEngineByName("GraalVM:js");
         
@@ -439,7 +386,7 @@ public final class GraalEnginesTest2 extends NbTestCase {
     }
 
     @Test
-    public void testAllAccessEnabledAttribute() throws Exception {
+    public void allAccessEnabledAttribute() throws Exception {
         ScriptEngineManager man = Scripting.newBuilder().build();
         ScriptEngine js = man.getEngineByName("GraalVM:js");
