+++ ide/libs.graalsdk/test/unit/src/org/netbeans/libs/graalsdk/ScriptingTutorialTest.java
@@ -40,9 +40,7 @@ public class ScriptingTutorialTest extends NbTestCase {
 
         return NbModuleSuite.emptyConfiguration()
             .gui(false)
-            .clusters("platform|webcommon|ide")
             .honorAutoloadEager(true)
-            .enableClasspathModules(false)
             .addTest(ScriptingTutorial.class)
             .suite();
     }
