+++ ide/libs.truffleapi/manifest.mf
@@ -4,5 +4,3 @@ OpenIDE-Module: org.netbeans.libs.truffleapi
 OpenIDE-Module-Localizing-Bundle: org/netbeans/libs/truffle/Bundle.properties
 OpenIDE-Module-Specification-Version: 1.20
 OpenIDE-Module-Provides: com.oracle.truffle.polyglot.PolyglotImpl
-OpenIDE-Module-Hide-Classpath-Packages: com.oracle.truffle.**,jdk.vm.ci.services.**
-
