+++ nbbuild/antsrc/org/netbeans/nbbuild/extlibs/ignored-overlaps
@@ -128,7 +128,3 @@ rust/rust.grammar/external/ST4-4.3.4.jar ide/libs.antlr4.runtime/external/ST4-4.
 
 # The enterprise usage is for library - it is better to keep the two separate
 ide/servletapi/external/jakarta.servlet-api-4.0.4.jar enterprise/websvc.restlib/external/jakarta.servlet-api-4.0.4.jar
-
-# The library is used for compilation, but is not distributed, as graalsdk.system relies on JDK-provided packages at runtime
-ide/libs.graalsdk/external/graal-sdk-20.3.0.jar ide/libs.graalsdk.system/external/graal-sdk-20.3.0.jar
-ide/libs.graalsdk/external/launcher-common-20.3.0.jar ide/libs.graalsdk.system/external/launcher-common-20.3.0.jar
