+++ nbbuild/cluster.properties
@@ -391,7 +391,6 @@ nb.cluster.ide=\
         libs.freemarker,\
         libs.git,\
         libs.graalsdk,\
-        libs.graalsdk.system,\
         libs.ini4j,\
         libs.jaxb,\
         libs.jcodings,\
