+++ platform/api.scripting/nbproject/project.xml
@@ -42,11 +42,6 @@
                         <recursive/>
                         <compile-dependency/>
                     </test-dependency>
-                    <test-dependency>
-                        <code-name-base>org.netbeans.libs.graaljs</code-name-base>
-                        <recursive/>
-                        <compile-dependency/>
-                    </test-dependency>
                 </test-type>
             </test-dependencies>
             <public-packages>
