+++ platform/api.scripting/test/unit/src/org/netbeans/api/scripting/ScriptingTutorialTest.java
@@ -19,7 +19,6 @@
 package org.netbeans.api.scripting;
 
 import javax.script.ScriptEngine;
-import static org.junit.Assume.assumeTrue;
 import org.netbeans.junit.NbTestCase;
 
 public class ScriptingTutorialTest extends NbTestCase {
@@ -30,7 +29,7 @@ public class ScriptingTutorialTest extends NbTestCase {
     public void testFourtyTwo() throws Exception {
         // @start region="testFourtyTwo"
         ScriptEngine js = Scripting.createManager().getEngineByMimeType("text/javascript");
-        assumeTrue(js != null || System.getProperty("java.vendor.version").contains("Graal"));
+        assert js != null;
 
         Number x = (Number) js.eval("6 * 7");
