+++ webcommon/libs.graaljs/manifest.mf
@@ -5,5 +5,3 @@ OpenIDE-Module-Layer: org/netbeans/libs/graaljs/layer.xml
 OpenIDE-Module-Localizing-Bundle: org/netbeans/libs/graaljs/Bundle.properties
 OpenIDE-Module-Specification-Version: 1.20
 OpenIDE-Module-Provides: javax.script.ScriptEngine.js,org.netbeans.libs.graaljs
-OpenIDE-Module-Hide-Classpath-Packages: com.oracle.truffle.js.scriptengine.**, 
-    com.oracle.js.parser.**, com.oracle.truffle.js.**
